Added generator functionality and Rails 2.1 compatibility.
Generator functionality: instead of having the whole of Goldberg work
as a generator, the code to populate public/ with assets and to put in
the appropriate default template is now invoked from the rake task
"goldberg:theme[name]".  "name" is optional and can be one of "ewnf",
"goldberg", "snooker" or "spoiled_brat".  This uses Rails' generator
code to move just the appropriate files.

Fixed various issues for Rails 2.1 compatibility, including making
class names explicit in associations and fixing migrations.